--- example/server/config.json
@@ -1,32 +1,0 @@
-{
-  "registry": {
-    "io.hadron.spaceTelescope": {
-      "http://sandbox.dev.aikon.com:3405/": {
-        "offer": {
-          "address": "0x9492d278f51d8b0ee5526977ffe6bbcc3ccf7152"
-        }
-      },
-      "https://us-central1-partner-hadron.cloudfunctions.net/hadron": {
-        "offer": {
-          "address": "0xb17e4d0295b12b3db4503b94231ca373304631b9"
-        }
-      },
-      "https://hadron.aikon.com": {
-        "offer": {
-          "address": "0xef06eda62ccc83d5cc7c99cb7ba036b0ade7b889"
-        }
-      }
-    }
-  },
-  "verifier": {
-    "publicKey": "-----BEGIN PUBLIC KEY-----\nMFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEU0rYoE6B58xW2nBw2/zO5Y/WtuLcCozrsbBZrfXI4Pr8PiAXjC02dONn90RmUswXxTUBKtQHYGSjAaDHSEvRDw==\n-----END PUBLIC KEY-----"
-  },
-  "services": {
-    "web3": {
-      "endpoint": "http://private-geth.aikon.com:15458"
-    },
-    "ipfs": {
-      "endpoint": "http://ipfs2-ext.dev.api.market:5001"
-    }
-  }
-}
