+++ lib/client.js
@@ -2,7 +2,6 @@ const keythereum = require('keythereum')
 const utils = require('ethereumjs-util')
 const { instrument, capability, contract } = require('open-rights-exchange')
 
-const { fromRegistry } = require('./server')
 const { connectIPFS } = require('./ipfs')
 const { connectWeb3 } = require('./web3')
 
@@ -47,12 +46,24 @@ exports.init = async (config, keys, _options={}) => {
 
   return {
     fetch: async (endpoint, request) => {
-      const voucherAddress = fromRegistry(config, endpoint).voucher.address
-      const apiCapability = await open(voucherAddress)
+      const apiCapability = await open(fromRegistry(config, endpoint).voucher.address)
 
       return capability.perform(apiCapability, request)
     }
   }
+}
+
+const fromRegistry = (config, endpoint) => {
+  const registry = config.registry
+  const names = Object.keys(registry)
+
+  if (names.length !== 1) {
+    throw new Error('only one offer is currently supported')
+  }
+
+  const key = names[0]
+  const scope = registry[key]
+  return scope[endpoint]
 }
 
 exports.purchase = async (offerAddress) => {
