+++ package.json
@@ -5,8 +5,7 @@
   "main": "./dist/index.js",
   "scripts": {
     "build": "babel . --ignore node_modules,dist --out-dir dist && npm run copyfiles",
-    "copyfiles": "cp ./example/client/config.json ./dist/example/client && cp ./example/server/config.json ./dist/example/server",
-    "example:server": "node --trace-warnings ./dist/example/server/index.js",
+    "copyfiles": "cp ./example/client/config.json ./dist/example/client",
     "example:client": "node --trace-warnings ./dist/example/client/index.js",
     "prepublish": "npm run build",
     "test": "echo \"Error: no test specified\" && exit 1"
